Accept new incoming connection in main 'thread'
This removes the need for the pipes between main and bridge code, so
they have been removed. It also removes the need for a flag used to
indicate if the main code should listen on the socket, thereby
simplifying things quite a bit.
It does mean that while accepting a connection that processing for other
modems will stop, but I don't see that as a big thing as accept() is
unlikely to block for any length of time, and it's most likely that this
program will only ever be used with a single modem.